Move water tool functions from TopToolbar.cpp into Water.cpp
